refactor: make Cipher's dimensions usize type

+++ exercises/_exercism/67_crypto-square/src/lib.rs
@@ -3,8 +3,8 @@ pub fn encrypt(input: &str) -> String {
 }
 
 struct Cipher {
-    rows: u64,
-    cols: u64,
+    rows: usize,
+    cols: usize,
     normal: Vec<char>,
 }
 
@@ -19,12 +19,12 @@ impl Cipher {
         Self { rows, cols, normal }
     }
 
-    fn determine_dimensions(normal_length: f64) -> (u64, u64) {
-        let maybe_perfect_sqrt = normal_length.sqrt().round() as u64;
-        if maybe_perfect_sqrt.pow(2) == normal_length as u64 {
+    fn determine_dimensions(normal_length: f64) -> (usize, usize) {
+        let maybe_perfect_sqrt = normal_length.sqrt().round() as usize;
+        if maybe_perfect_sqrt.pow(2) == normal_length as usize {
             (maybe_perfect_sqrt, maybe_perfect_sqrt)
         } else {
-            let c = (0.25 + normal_length.sqrt() + 0.5).ceil() as u64;
+            let c = (0.25 + normal_length.sqrt() + 0.5).ceil() as usize;
             (c - 1, c)
         }
     }
